Update lemon-squeezy.md - Round #1
Here's the changes I made to the first script. It might be a lengthy video, but I think it's appropriate since we're establishing what we're doing, why, and we're making our first product to put on the storefront. We basically have one goal in this episode, so I don't think it's necessary to chop it up into two episodes.