Update checkout-api.md - Round #1
You'll definitely want to check this one for technical accuracy. There were a few moments where I wasn't 100% sure what you meant, so I did my best to figure it out. I reworded/added some lines based on what I saw in the video portion. And I left the linked terms, like "[Create a checkout]", in place.

As a side note, the video, even with the cuts you already made, is almost 29 minutes long. Is that the one we're going with for the final edit? If so, that's going to take me a while to edit down to what I imagine will be no more than maybe 6-8 minutes of accompanying audio. If a re-record isn't too much trouble, especially using the final version of the script so it's an even closer match, that would be much easier to work with.