Update multiple-products-purchase.md - Round #1
I made A LOT of changes here, so please make sure the content is still accurate. I cut several sentences because I felt they were either unnecessary or too repetitive. Around line 9, I cut a whole segment because it seemed like we talked about doing something that we didn't actually do (I think?).

I also cut out the segment near the end where we hit "continue" and then just kind of end up in the same place. It didn't seem necessary to include that when we were so close to the end.

Hopefully all of that makes sense. This one could potentially benefit from a re-record with the new script in mind, if you guys find that the cuts I made are useful/appropriate. I'll leave that up to you guys.